Fixed Best() operator. The EV output is now consistent with Simulation. I believe both are working correctly now. Standard algorithm converges as expected.

+++ Poker/Node/FlopNode.cs
@@ -32,7 +32,7 @@ namespace Poker
         public override void CalculateBest()
         {
             //return;
-            float SingleTurnWeight = 1f / (Card.N - 4 - 3);
+            double SingleTurnWeight = 1f / (Card.N - 4 - 3);
             CalculateBest_AccountForOverlaps(SingleTurnWeight);
         }
